Disable check-doc.py test in Travis
Summary:
This script is broken due to major reformatting changes which
it does not understand yet.

Since it's not helping by breaking Travis builds, disable it
for now.

Test Plan:
Check that Travis does not fail with a message like this one:

`The command "if [ "$CHECK_DOC" = 1 ]; then contrib/devtools/check-doc.py; fi" failed and exited with 73 during .`

Reviewers: deadalnix, #bitcoin_abc

Reviewed By: deadalnix, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D69