Fix blockcheck_tests initial conditions (set max block size to default)
Summary:
This test was failing because another had left the config with a max block
size that was not default.

Each test has to ensure its own appropriate initial conditions.

Test Plan:
  make check

See T19 - it is necessary that we run 'test_bitcoind' with different random seeds from time to time (iterative testing for a certain time, to see if it fails on certain orderings of tests.) Another good reason to get our own CI up and running ASAP.

Reviewers: #bitcoin_abc, deadalnix, sickpig, kyuupichan

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D146