Change default value for fork MG from 8 to 2MB
Summary: This is safer default. 8MB is close to an order of magnitude larger than current block size. If something doesn't quite scale as expected, we need to see it comming be able to adapt. This imply doing a smooth rollout while monitoring. A default MG of 2MB ensure that goal.

Test Plan:
  \o/

Reviewers: freetrader, sickpig, kyuupichan, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D152