Separate the test plan from the requirements specification
Summary:
The test plan is separated out from the spec, but it still needs to
be revised. All review comments and suggestions for additions/filling
gaps are hugely welcomed!

The reference URL [4] introduced in the req spec needs to be finalized.
It is currently pointing to the Github master which does not have the
latest source and docs.

This Diff the equivalent of BUIP PR 51 / issue 50 , so I repeat the reasoning:

- requirements spec needs to be kept short
- test plan will grow and is client specific

Test Plan: read andreview

Reviewers: deadalnix, sickpig, kyuupichan, Mengerian, Juan.Bitprim, #bitcoin_abc

Reviewed By: deadalnix, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D177