Update doc/release-process.md for ABC
Summary:
This is intended to create a more useful guideline for helping with
Bitcoin ABC releases.

It is fairly minimal for now. The goal is to help provide a checklist
to make releases easier.

This document will need to be expanded and refined over time.

This document was created in conjunction with freetrader

Test Plan: Read the doc as we do a release

Reviewers: freetrader, deadalnix, O1 Bitcoin ABC, #bitcoin_abc

Reviewed By: freetrader, deadalnix, O1 Bitcoin ABC, #bitcoin_abc

Subscribers: sickpig

Differential Revision: https://reviews.bitcoinabc.org/D335