Remove version from UTXO and undo datastructures.
Summary:
This is a continuation of the work started with D342 . As this info is not very useful, we don't want to be saving it for every single UTXO so we drop it.

This require to add a specific way to verify the hash for on disk data as some of the legacy data do contain a version number that we do not parse anymore, and, as a result, checksum do not match the old way.

Test Plan:
  make check
  ../../qa/pull-tester/rpc-tests.py

Start a node on a legacy .bitcoin folder, verify that it pick it up and can sync from it.
Perform a reorg and verify that it behaves properly.

Reviewers: freetrader, kyuupichan, awemany, sickpig, #bitcoin_abc, CCulianu

Reviewed By: freetrader, awemany, #bitcoin_abc, CCulianu

Subscribers: zander, CCulianu

Differential Revision: https://reviews.bitcoinabc.org/D362