Reject non replay protected transactions after the fork.
Summary:
This ensure that the HF activates STRICTENCODING (just standard right now, not consensus) and it'll reject any transaction not using SIGHASH_FORKID after the fork.
To make sure we get a clean mempool, the mempool is cleared when we connect a block that activate the fork.

Numerous places have to check for transactionw ith and without the forkid flag. This is abit wonky, but, hopefully, we'll be able to cleanup that mess after Aug, 1 and that'll come very soon.

Test Plan:
Updated related tests.

  make check
  ../qa/pull-tester/rpc-tests.py -extended

Reviewers: freetrader, sickpig, kyuupichan, #bitcoin_abc

Reviewed By: freetrader, kyuupichan, #bitcoin_abc

Subscribers: zander

Differential Revision: https://reviews.bitcoinabc.org/D371