Fix small Mac OSX memory leak (imported from BU sources)
Summary:
I noticed there was a very small memory leak in the way CoreFoundation objects were being used on Mac OSX.  I submitted that patch months ago to the BU team to fix it, and it's been accepted and merged into their repository.  I wanted to bring that patch here to our project, just as a small nit, because it bugs me the way Core originally wrote this code.  They clearly don't know how to use the CoreFoundation API ;)

Here is the reference in the BU repository:

https://github.com/BitcoinUnlimited/BitcoinUnlimited/pull/507/commits/045a999785d213a2366158fc4372f51041800e7a

Test Plan: Compile on OSX if you dare, open Qt gui app, toggle load on startup on/off

Reviewers: deadalnix, sickpig, freetrader, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Subscribers: ealmansi

Differential Revision: https://reviews.bitcoinabc.org/D452