Fix RPC signrawtransaction silently accepting missing amount field, make it throw an error instead
Summary:
After hard fork, replay protected tx's require the amount field to be non-zero for previous outputs in signed tx's. JSON RPC code was accepting transactions with this field missing when it should have been throwing an error. Users had the impression the transaction that was signed was good, but when trying to sent to network, they would get an error that the tx is bad. See github issue https://github.com/Bitcoin-ABC/bitcoin-abc/issues/63.

Depends on D448

Test Plan:
Depends on D448

make check and also rpc-tests.py

Reviewers: #bitcoin_abc, sickpig, freetrader, dgenr8, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D449