Add the API to add/spend coin on a per UTXO basis.
Summary:
This ensure we can add/spend coin per utxo. It remove the ModifyNewCoins new API and makes the ModifyCoins on obsolete. Once the API is fully migrated, the storage can move to a work on a per UTXO basis.

This is a continuation of the work started with D342 .

Test Plan:
Updated relevent tests.

  make check
  ../qa/pull-tester/rpc-tests.py

Reviewers: freetrader, CCulianu, #bitcoin_abc

Reviewed By: CCulianu, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D470