Tweak to travis.yml to exit on error
Summary: Making travis exit on error during a certain part of the build is a good thing.  Added set -e to a part of the travis script to just die on build error.  Previously travis would just happily continue even if there were errors in the executed commands.

Test Plan: Push a commit to github with travis enabled and see it work.  If you want, push a commit with a build error to see it die right after the compile error (rather than try and run the tests anyway as was the previous behavior).

Reviewers: freetrader, deadalnix, #bitcoin_abc

Reviewed By: freetrader, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D492