Fix proxy_test logging crash when no IPv6 available
Summary:
The test tried to use the self.log member in `__init__()`
but it is only set up in main() in BitcoinTestFramework.

This moves the calls into setup_nodes() which is called
by main, and should log correctly if IPv6 is not available.

Test Plan:
  qa/pull-tester/rpc-tests.py proxy_test
check that the test no longer fails on Travis

Reviewers: #bitcoin_abc, deadalnix, CCulianu

Reviewed By: #bitcoin_abc, deadalnix, CCulianu

Differential Revision: https://reviews.bitcoinabc.org/D504