Fix PYTHONPATH export association in Makefile.am
Summary:
This fixes a build problem on FreeBSD which was raised by /u/1s44c in [1].
The problem with the original code that the "export PYTHONPATH" is
interpreted by gmake as an improperly indented command for the preceding
target.

Indentation, while it would surmounting the immediate build problem on FreeBSD,
is not the correct solution, since the export is intended to apply to further
command sequences and not to be a associated to the particular target
listed above it.

Accordingly, it is moved up which has been verified by inspection to result
in a clean Makefile on FreeBSD and Linux.

[1] https://np.reddit.com/r/BitcoinABC/comments/6yuy53/bitcoin_abc_ubuntu_repo_has_been_updated_to_serve/dms6ip3/

Test Plan:
Test fix on FreeBSD by building from scratch (from autogen.sh thru rpc-tests)
Check Travis builds are all OK
Check Gitian builds are all OK
(if possible: OSX native build from scratch)

Reviewers: #bitcoin_abc, deadalnix, CCulianu

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D543