Display if a node is using the cash magic in getpeerinfo
Summary: As per title. This will be useful to know when it is time to disable the legacy magic in the network layer.

Test Plan:
Run a node on testnet, call getpeerinfo, ensure presented data are correct.
Modify a node to connect with the new cash magic and ensure it is properly marked as using cash.

Reviewers: freetrader, CCulianu, dagurval, #bitcoin_abc, schancel

Reviewed By: #bitcoin_abc, schancel

Subscribers: schancel

Differential Revision: https://reviews.bitcoinabc.org/D546