Make bitcoin-seeder a first class citizen of ABC
Summary:
Bitcoin seeder is a tool make by Pieter Wuille that is useful to create seeders. In order to reduce the maintenance burden, I wish to add it as a firstc lass citizen in ABC. This way, it'll be able to share code with ABC - which should provide good insight on what needs to be refactored and how - and can be distributed with it.

The seeder has been modified to use the Cash Magic when etablishing connexions and filter nodes based on NODE_BITCOIN_CASH rather than NODE_WITNESS which doesn't exist ont he Bitcoin Cash network. This work is based on the Bitcoin Cash seeder by @sickpig .

It comes with integration in the current build system.

Test Plan:
  make check

Ensure that the bitcoin-seeder executable is created properly. Run it and see it behaves as expected.

Reviewers: sickpig, schancel, dagurval, freetrader, #bitcoin_abc

Reviewed By: sickpig, schancel, #bitcoin_abc

Subscribers: sickpig

Differential Revision: https://reviews.bitcoinabc.org/D559