Cashaddr decoding the raw data and checksum
Summary:
This is split from D544 plus implement a candidate BCH code of degree 8 for it. The BCH code can detect up to 6 errors over the typical length of an address and up to 8 errors in a row.

This is intended to serve as a base to get the new address format started.

Test Plan: Added unittests.

Reviewers: #bitcoin_abc, dagurval

Reviewed By: #bitcoin_abc, dagurval

Subscribers: dagurval

Differential Revision: https://reviews.bitcoinabc.org/D576