Modify QR code to fit cashaddr
Summary:
This bumps the size of the QR code and also decreases the font size when
using cashaddr.

The encoded address is displayed as text in the QR code image. This allows
for the longer cashaddr to fit in the image.

Test Plan: Create a request using QT client on mainnet. Tested on Ubuntu.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D802