Update Berkeley DB to 5.3 minimum
Summary:
Berkeley DB changed it's internal format after 4.8 . So any DB created with 5.3 cannot be read with software using 4.8 . Historically, Bitcoin ABC ensure that the version used is 4.8 unless specified otherwise. This is cumbersome as this version is relatively old. In addition, it is fairly common that bitcoin is packaged using a more recent version by default (for instance on ubuntu/debian).

As a result, it seems like a better strategy to ensure the version used is 5.3 or more recent going forward.

Test Plan: Do gitian builds.

Reviewers: #bitcoin_abc, schancel

Reviewed By: #bitcoin_abc, schancel

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1330