Updated to allow macOS build
Summary: The existing Bitcoin ABC code would not compile on macOS.  It appears to be related to Boost.  Adding "const" to a few lines in miner.h and txmempool.h allowed the build process to proceed.  This change is consistent with a change performed on Bitcoin Core (https://github.com/bitcoin/bitcoin/pull/12032/files).  Make boost::multi_index comparators const.

Test Plan: Confirm the changes allow the build process to proceed as desired on Linux/Windows

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix, BMahon9

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D855