[qa] replace specific unit test checks with `make check`
Summary:
We were previously running unit tests for each package manually.
This was because we wanted JUNIT format results from the boost tests.
However, we had to downgrade the build system for the time being to
a version of boost which does not support JUNIT output.

Here, we switch to runnning `make check` as there is no benefit to running
each test suite individually

Test Plan:
  make check

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix, jasonbcox

Subscribers: teamcity

Differential Revision: https://reviews.bitcoinabc.org/D1422