Fix errors in maximum block size check.
We have a check to ensure that the configured excessive block size
does not exceed the maximum block file size, because if it does bitcoind
gets into an infinite loop and runs out of memory. There were 2
problems with this check however:

1) It had an off by 1 error; The actual problem in bitcoind is that
in validation.cpp FindBlockPos() it checks to see if the used space
in the current block file + the new block size is >= the maximum
block file size, then it allocates a new block file. In our new check
however we used >, not >= to reject large configured block sizes and
so would be off by 1 on what was actually required.

I actually think that check is wrong and it should be > instead of >=
(if the current file uses 0 bytes, the max size is X bytes, why would
we not put an X byte block into it?). But to fix that requires
careful checking and we don't have time for that right now.

2) I hadn't noticed that when bitcoin writes a block to the block file
it actually includes a fixed 8 byte header along with every block. So
the maximum block size we can write to a block file is actually 8 bytes
less then we initially thought.