Allow removal of txns from the transaction propagation logic.
If we get a new block, some of the transactions from there may
currently be queued for sending to our peers and it will be
a waste of time and bandwidth to continue sending them out.

Provide the ability to prune those transactions from our peers
pending transaction lists at the same time as they are pruned
from the mempool.