remove max ecessive size relation to max block file size
Max excessive block size becomes unlimited with this commit.
This produces the possibility that max block file size is less than excessive
block size so we have to handle it in get file position function so that it
simply writes a large block to a separate file and not cause infinite loop.