Improve stability of pruning functional test.
As part of the pruning.py test it resends all the wallet
transactions back into node0. This then relays those txns
on to node1. Node1 is run with mempool checking enabled,
which makes the processing of those received txns very slow,
and can hold up the processing of the blocks that are
required for the test to pass.

During the parts of the test most affected by the slow txn
processing, add the flag -checkmempool=0 to node1 to speed
things up.