Classes to support journal tracking of changes to the mempool.
We want to track changes to the mempool so support more efficient
block mining candidate construction. This commit introduces the
basic classes to help support this:

CJournal: A class to contain an ordered list of transactions for
inclusion in teh next mining candidate:

CJournalChangeSet: A list of changes to apply in a single batch
to a journal.

CJournalBuilder: A class for managing the current journal.