Release cs_main lock during script validation
Blocks are validated in parallel by releasing cs_main
lock during script validation - the part that takes up
most block processing time. In case a different block
with more or the same amount of work is validated before
the block that started validation first, the first block
is not attached as the new tip but rather only marked
as validated while the tip switch code is skipped.