Add flag to ContextualCheckTransaction to indicate if txn is from a block.
If the txn is non-final and came from a block then it should be a hard
error. If the txn wasn't from a block then we might want to still
accept it into the non-final pool.

Add an extra flag to ContextualCheckTransactionso that we can tell whether
to fail hard.