Check if we are on a production tag before enabling message.
If the --enable-prod-build configure option is passed we disable
the printing of the pre-release warning message. But we only
want to disable the message if we are not building from a beta
tag (or other non-production tag).

This sommit includes a script that checks if we are on a
production tag of the form vX.Y.Z and only disables the warning
message if so.