Peer average bandwidth measurement.
Measure and keep a record of the average download bandwidth for each
of our peers. This will be later used to help detect stalling and
other slow peer related issues.

We keep 2 measurements; a spot rate measured over a short 5 second
period, and a longer 1 minute rate measured over a moving window.