Orphan tracking should hold weak_ptr to node.
By holding a full shared_ptr to the node the orphan txn came from,
the orphan transaction tracking is inadvertently keeping that CNode
object alive beyond the time when the P2P networking layer has
decided we should close the connection to that node.

It is not until the time when the CNode object is reaped that any
blocks we were trying to download from that peer are added back to the
list of outstanding blocks and re-download attempts are made from
other peers. So by keeping the CNode alive the orphan tracking is
actually seriously delaying attempts to re-download failed blocks.

Fix this by holding a weak_ptr instead of a shared_ptr.