Make p2p timeouts configurable.
Due to head-of-line blocking within TCP, a very large message (such
as a block or a transaction) can hold up another message that is queued
from the same peer.

For example; we request a large block from a node and then send a ping
request to that node. That peer starts to send us the large block and
will not then be able to send us the ping response until the block transfer
has finished. We may timeout on the ping response and disconnect that
peer even though it was still active.

This problem will only manifest if the bandwidth between the nodes is not
sufficient to complete the big block / transaction within the timeout
interval, but now we have some very large blocks in the STN it can hit,
particularly when testing with artifically restricted bandwidth.

The correct solution is to fix the P2P stack so that small messages
such as pings are sent over a different connection to large block
and transaction messages, thus avoiding the head-of-line blocking
altogether. Until we can do that however just make the timeouts within
the P2P configurable so that we can work around it in testing if
required.