Increase size of setAskFor.
CNode::setAskFor was twice the size of mapAskFor, this makes it four
times the size. Note that mapAskFor entries are cleared out at the
point we request the txn, but setAskFor entries are not removed
until we actually receive the requested txn. If there is a large
latency between requests and responses then setAskFor can quickly
become full and overflow.

From observation on the STN, a significant cause of txns being lost is
setAskFor becoming full due to the large numbers of incoming transactions
and the time it takes for some of those txns to actually arrive after
being requested. Increasing the size of setAskFor should help the
situation.