Remove unnecessary thread checks before deleting CNode.
The checks in CConnman::ThreadSocketHandler to see if any threads
are still using a CNode before deleting it are unnecessary and
confusing, so remove them.

Now we are using a shared pointer to CNode instead of a raw CNode*,
we can simply use the use count on the shared pointer to safely
and simply check whether there are any users of that CNode still
in existance.

I have audited every code path that takes the 2 mutexes previously
used as an extra check before deletion, and in all cases any references
held by those threads to the CNode were copies of the master CNodePtr
held in CConnman::vNodes made while under the protection of cs_vNodes.
Therefore using the reference count from within the shared pointer is
just as safe as the previous double mutex lock check, and much cleaner.