Fix P2P data xfer deadlock.
It has been observed that sometimes if 2 nodes are both trying to exchange
large amounts of data at the same time they can both actually completely
stop sending, and data transfer entirely stops between them.

This is due to the `continue` statement at line 1533 in net.cpp. This has
the effect of preventing a node trying to receive any data until it has no
data queued for sending, but if the other node is also doing the same thing
then they can both fill up their socket buffers and will then neither be
sending or receiving.

Remove that `continue` statement so we do proper read/write multiplexing
on the sockets.