Increase rate at which JBA processes new transactions.
There's a lot of contention for cs_main, which the JBA needs to
take each time its background thread runs. This has the effect of
reducing the rate at which the JBA actually builds the next candidate
block from the journal.

Increase the rate at which we pull transactions out of the journal
to make better use of each time slot we run.