Optimization for removing for block and adding for reorg
When transactions were removed from the mempool, because of a new block is mined, they were checked one by one, which in case of the long chains, when we were removing parent we were checking children which would be removed in the next step. now we are checking only transactions that stay in the mempool.
In the case of reorg we were adding transactions one by one and checking them, but on the end of the process we were checking them (with theirs children) again. now at the end of the process we are checking their children only.